# mate-platform-backend/contracts/openapi/generated/bundled.yaml
openapi: 3.1.0
info:
  title: Mate Platform API
  version: 1.0.0
  description: Canonical contract-first API for all Mate Platform domains.
  contact:
    name: Mate Platform API Governance
    email: api-governance@metaplatform.local
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-mate-owner: architecture-api-governance
servers:
  - url: /
    description: Gateway-relative production endpoint
security:
  - bearerAuth: []
...
paths:
  /api/v1/a2a/agent-cards/search:
    get:
      tags:
        - a2a
      summary: /v1/a2a/agent-cards/search
      operationId: a2aGetA2aAgentCardsSearch
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
      description: /v1/a2a/agent-cards/search
      x-mate-owner: ai-protocols
      x-mate-permission: a2a.get
      x-mate-requirements:
        - FR-A2A-A2AGETA2AAGENTCARDSSEARCH
      x-mate-implementation-status: planned
  /api/v1/a2a/delegations:
    get:
      tags:
        - a2a
      summary: /v1/a2a/delegations
      operationId: a2aGetA2aDelegations
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
      description: /v1/a2a/delegations
      x-mate-owner: ai-protocols
      x-mate-permission: a2a.get
      x-mate-requirements:
        - FR-A2A-A2AGETA2ADELEGATIONS
      x-mate-implementation-status: planned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    ApiResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          enum:
            - 0
        message:
          type: string
        data: {}
        traceId:
          type: string
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
        traceId:
          type: string
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: 服务器错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...